Prep for wiegand data processing
