+++ src/main/java/com/ams/Autograder/Autograder.java
@@ -1,6 +1,6 @@
 package com.ams.autograder;
 
-import com.ams.restapi.attendance.AttendanceLog;
+import com.ams.restapi.attendance.AttendanceRecord;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -51,7 +51,7 @@ public class Autograder {
         submissionsOptions.setStudentSubmissionOptionMap(mapOfOptions);
         submissionWriter.gradeMultipleSubmissionsByCourse(submissionsOptions);
 
-        HashMap<String, List<AttendanceLog>> attendanceMap = new HashMap<>();
+        HashMap<String, List<AttendanceRecord>> attendanceMap = new HashMap<>();
         
 
         //listing every user inside a course
