+++ src/main/java/com/ams/restapi/attendance/AttendanceRecord.java
@@ -1,0 +1,146 @@
+package com.ams.restapi.attendance;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+
+@Entity
+public class AttendanceRecord {
+
+    public static enum AttendanceType {
+        ARRIVED, ARRIVED_LATE, ARRIVED_INVALID, LEFT_INVALID, LEFT, INVALID
+    }
+
+    private @Id @GeneratedValue Long id;
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    private String room;
+    private LocalDate date;
+    private LocalTime time;
+    private String sid;
+    private String type; // I'll be the judge of this
+    
+    public String getRoom() {
+        return room;
+    }
+
+    public void setRoom(String room) {
+        this.room = room;
+    }
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public void setDate(LocalDate date) {
+        this.date = date;
+    }
+
+    public LocalTime getTime() {
+        return time;
+    }
+
+    public void setTime(LocalTime time) {
+        this.time = time;
+    }
+
+    public String getSid() {
+        return sid;
+    }
+
+    public void setSid(String sid) {
+        this.sid = sid;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    AttendanceRecord() {}
+
+    public AttendanceRecord(String room, LocalDate date,
+        LocalTime time, String sid, String type) {
+            this.room = room;
+            this.date = date;
+            this.time = time;
+            this.sid = sid;
+            this.type = type;
+    }
+    
+    @Override
+    public String toString() {
+        return "AttendanceLog [id=" + id + ", room=" + room + ", date=" + date + ", time=" + time + ", sid=" + sid
+                + ", type=" + type + "]";
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((id == null) ? 0 : id.hashCode());
+        result = prime * result + ((room == null) ? 0 : room.hashCode());
+        result = prime * result + ((date == null) ? 0 : date.hashCode());
+        result = prime * result + ((time == null) ? 0 : time.hashCode());
+        result = prime * result + ((sid == null) ? 0 : sid.hashCode());
+        result = prime * result + ((type == null) ? 0 : type.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        AttendanceRecord other = (AttendanceRecord) obj;
+        if (id == null) {
+            if (other.id != null)
+                return false;
+        } else if (!id.equals(other.id))
+            return false;
+        if (room == null) {
+            if (other.room != null)
+                return false;
+        } else if (!room.equals(other.room))
+            return false;
+        if (date == null) {
+            if (other.date != null)
+                return false;
+        } else if (!date.equals(other.date))
+            return false;
+        if (time == null) {
+            if (other.time != null)
+                return false;
+        } else if (!time.equals(other.time))
+            return false;
+        if (sid == null) {
+            if (other.sid != null)
+                return false;
+        } else if (!sid.equals(other.sid))
+            return false;
+        if (type == null) {
+            if (other.type != null)
+                return false;
+        } else if (!type.equals(other.type))
+            return false;
+        return true;
+    }
+
+    
+
+}
