+++ src/main/java/com/ams/restapi/courseInfo/CourseInfoRespository.java
@@ -1,7 +1,9 @@
 package com.ams.restapi.courseInfo;
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CourseInfoRespository extends JpaRepository<CourseInfo, Long> {
-
+    // public Optional<CourseInfo> findByRoomAndTimeBetween
 }
